LayoutTests/fast/dom/StyleSheet/gc-styleheet-wrapper.xhtml

<?xml-stylesheet href="resources/detached-style.css" title="testSheet"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title/>
<link rel="stylesheet" href="resources/detached-style.css" type="text/css" title="testSheet"/>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<style title="testSheet">
* { }
</style>
</head>
<body onload="start()">
<script>
    window.jsTestIsAsync = true;

    description("Test that custom properties on stylesheets survive GC");

    function test(expr)
    {
        shouldBe(expr + ".title", "'testSheet'");

        eval(expr).foo = "bar";

        gc();

        shouldBe(expr + ".foo", "'bar'");
    }

    function start()
    {
        test("document.firstChild.sheet");
        test("document.getElementsByTagName('link')[0].sheet");
        test("document.getElementsByTagName('style')[0].sheet");
        finishJSTest();
    }

    successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
