third_party/WebKit/LayoutTests/fast/dom/StyleSheet/gc-styleheet-wrapper-expected.txt

Test that custom properties on stylesheets survive GC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.firstChild.sheet.title is 'testSheet'
PASS document.firstChild.sheet.foo is 'bar'
PASS document.getElementsByTagName('link')[0].sheet.title is 'testSheet'
PASS document.getElementsByTagName('link')[0].sheet.foo is 'bar'
PASS document.getElementsByTagName('style')[0].sheet.title is 'testSheet'
PASS document.getElementsByTagName('style')[0].sheet.foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
